+++ packages/signers/pjs-signer/src/from-pjs-account.ts
@@ -5,8 +5,7 @@ import {
   type V15,
   compact,
   decAnyMetadata,
-  enhanceEncoder,
-  u8,
+  extrinsicFormat,
 } from "@polkadot-api/substrate-bindings"
 import { fromHex, mergeUint8, toHex } from "@polkadot-api/utils"
 import { getDynamicBuilder, getLookupFn } from "@polkadot-api/metadata-builders"
@@ -23,12 +22,6 @@ export const getAddressFormat = (metadata: V14 | V15): number => {
 
   return dynamicBuilder.buildDefinition(constant.type).dec(constant.value)
 }
-
-const versionCodec = enhanceEncoder(
-  u8.enc,
-  (value: { signed: boolean; version: number }) =>
-    (+!!value.signed << 7) | value.version,
-)
 
 const getPublicKey = AccountId().enc
 export function getPolkadotSignerFromPjs(
@@ -71,6 +64,8 @@ export function getPolkadotSignerFromPjs(
     pjs.signedExtensions = []
 
     const { version } = decMeta.extrinsic
+    if (version !== 4 && version !== 5)
+      throw new Error(`Unsupported extrinsic version ${version}`)
     const extra: Array<Uint8Array> = []
 
     decMeta.extrinsic.signedExtensions.map(({ identifier }) => {
@@ -110,10 +105,13 @@ export function getPolkadotSignerFromPjs(
 
     if (!result.signedTransaction) {
       const preResult = mergeUint8(
-        versionCodec({ signed: true, version }),
+        extrinsicFormat[0]({ version, type: "signed" }),
         // converting it to a `MultiAddress` enum, where the index 0 is `Id(AccountId)`
         new Uint8Array([0, ...publicKey]),
         fromHex(result.signature),
+        // FIXME: extension version should come from metadata, already
+        // follow https://github.com/paritytech/polkadot-sdk/pull/3685
+        version === 5 ? new Uint8Array([0]) : new Uint8Array(),
         ...extra,
         callData,
       )